safari-534.53-branch/LayoutTests/platform/mac/accessibility/select-element-selection-with-optgroups-expected.txt


This tests that setting selection within a list box works correctly if there are optgroups

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS selectElement.selectedChildrenCount is 0
PASS selectElement.selectedChildrenCount is 1
PASS selectElement.selectedChildAtIndex(0).isEqual(option1) is true
PASS selectElement.selectedChildrenCount is 1
PASS selectElement.selectedChildAtIndex(0).isEqual(option2) is true
PASS selectElement.selectedChildrenCount is 1
PASS selectElement.selectedChildAtIndex(0).isEqual(option3) is true
PASS successfullyParsed is true

TEST COMPLETE
